+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* added `ScanQueryMode` for `Table::scanQuery`
+
 ## 1.13.2
 ### Bugs